requirements.txt
requests>=2.25.1
pillow>=8.0.0
numpy>=1.21.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.0
pyyaml>=5.4.0
python-dateutil>=2.8.2
mahjong>=1.2.0
